Refactor label
